Fixed attack bug

+++ src/characters/fighter.rs
@@ -11,8 +11,8 @@ pub trait Fighter {
     fn attack(&mut self, target: &mut dyn Fighter) -> String;
     fn take_damage(&mut self, damage: u16);
     fn attack_speed_as_milliseconds(&self) -> u16;
-    fn can_attack(&self, current_time: i32, last_attack_time: i32) -> bool {
-        current_time - last_attack_time >= self.attack_speed_as_milliseconds() as i32
+    fn can_attack(&self, current_time: u64, last_attack_time: u64) -> bool {
+        current_time - last_attack_time >= self.attack_speed_as_milliseconds() as u64
     }
 
     fn power(&self) -> u16;
